refactor(fees): externalize loading fees to own file
